Fixes #6
