+++ src/outputs/timekey/index.js
@@ -1,5 +1,7 @@
 import React from 'react';
 
+import Layout from './Layout'
+
 import Grid from '@material-ui/core/Grid';
 import { makeStyles } from '@material-ui/core/styles';
 import AppBar from '@material-ui/core/AppBar';
@@ -35,79 +37,11 @@ const useStyles = makeStyles(theme => ({
   table: {
     minWidth: 700,
   },
-  heading: {
-    fontSize: theme.typography.pxToRem(15),
-    flexBasis: '33.33%',
-    flexShrink: 0,
-  },
-  secondaryHeading: {
-    fontSize: theme.typography.pxToRem(15),
-    color: theme.palette.text.secondary,
-  },
 }));
 
-function ControlledExpansionPanels({date, billableHours, nonBillableHours, totalHours, children}) {
-
-  const classes = useStyles();
-  const [expanded, setExpanded] = React.useState(false);
-
-  const handleChange = panel => (event, isExpanded) => {
-    setExpanded(isExpanded ? panel : false);
-  };
-
-  return (
-      <ExpansionPanel expanded={expanded === 'panel1'} onChange={handleChange('panel1')}>
-        <ExpansionPanelSummary
-          expandIcon={<ExpandMoreIcon />}
-          aria-controls="panel1bh-content"
-          id="panel1bh-header"
-        >
-          <TableCell rowSpan={3} />
-          <TableCell className={classes.heading} colSpan={2}>{date}</TableCell>
-          <TableCell className={classes.secondaryHeading}>{billableHours}</TableCell>
-          <TableCell className={classes.secondaryHeading}>{nonBillableHours}</TableCell>
-          <TableCell className={classes.secondaryHeading}>{totalHours}</TableCell>
-        </ExpansionPanelSummary>
-        <ExpansionPanelDetails>
-          {children}
-        </ExpansionPanelDetails>
-      </ExpansionPanel>
-  );
-}
-
 function App() {
-  
-    const classes = useStyles();
-    const [value, setValue] = React.useState('recents');
-  
-    const handleChange = (event, newValue) => {
-      setValue(newValue);
-    };
-
-    function a11yProps(index) {
-      return {
-        id: `action-tab-${index}`,
-        'aria-controls': `action-tabpanel-${index}`,
-      };
-    }
-
   return (
-    <Grid container spacing={3}>
-      <AppBar position="static" color="default">
-        <Tabs
-          value={value}
-          onChange={handleChange}
-          indicatorColor="primary"
-          textColor="primary"
-          variant="fullWidth"
-          aria-label="action tabs example"
-        >
-          <Tab label="Item One" {...a11yProps(0)} />
-          <Tab label="Item Two" {...a11yProps(1)} />
-          <Tab label="Item Three" {...a11yProps(2)} />
-        </Tabs>
-      </AppBar>
-      <Grid item md={12} xs={12}>
+      <Layout>
         <TableContainer component={Paper}>
           <Table className={classes.table} aria-label="spanning table">
             <TableHead>
@@ -136,32 +70,7 @@ function App() {
             </TableBody>
           </Table>
         </TableContainer>
-      </Grid>
-      <Grid container item spacing={4} xs={12}>
-        <Grid item container md={3} xs={6}>
-          <Grid item xs md={6}>
-            <Paper>
-              <Button>Button1</Button>
-            </Paper>
-          </Grid>
-          <Grid item xs md={6}>
-          <Paper>
-            <Button>Button2</Button>
-          </Paper>
-          </Grid>
-        </Grid>
-        <Grid item md={6} xs={12} />
-        <Grid item md={3} xs={6}>
-          <Paper>
-            <Button>Button3</Button>
-          </Paper>
-        </Grid>
-      </Grid>
-      <Fab variant="extended" float="right">
-        <NavigationIcon className={classes.extendedIcon} />
-          Navigate
-      </Fab>
-    </Grid>    
+    </Layout>    
   );
 }
